requirements.txt
streamlit
bs4
BeautifulSoup
requests
pandas
textblob
TextBlob
wordcloud 
matplotlib
pyplot
urllib.parse
urljoin
spacy
json
sklearn.feature_extraction.text
TfidfVectorizer
collections 
Counter